+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Implemented references for the LSP server
 - Experiment with an MCP server
+- Fixed array equality
 
 ## [0.6.1] 2026-03-11